+++ packages/extension-polkagate/src/fullscreen/stake/stakingPositions/PositionItem.tsx
@@ -1,10 +1,10 @@
 // Copyright 2019-2025 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-//@ts-ignore
+// @ts-ignore
 import type { PalletNominationPoolsClaimPermission } from '@polkadot/types/lookup';
 
-import { Container, Grid, Stack, Typography } from '@mui/material';
+import { Box, Container, Grid, Stack, Typography } from '@mui/material';
 import { ArrowRight2, InfoCircle } from 'iconsax-react';
 import React, { memo, useCallback, useMemo } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
@@ -87,16 +87,17 @@ export const ChainIdentifier = ({ genesisHash }: TokenInfoProps) => {
 };
 
 interface Props extends TokenInfoProps {
-  type: 'pool' | 'solo';
   balance: BN;
   claimPermissions?: PalletNominationPoolsClaimPermission['type'],
-  price: number;
   decimal: number;
-  token: string;
   isSelected?: boolean;
+  price: number;
+  token: string;
+  totalPositions: number;
+  type: 'pool' | 'solo';
 }
 
-function PositionItem ({ balance, claimPermissions, decimal, genesisHash, isSelected, price, token, type }: Props) {
+function PositionItem ({ balance, claimPermissions, decimal, genesisHash, isSelected, price, token, totalPositions, type }: Props) {
   const { address } = useParams<{ address: string }>();
   const navigate = useNavigate();
   const hasPoolStaking = useMemo(() => type === 'pool', [type]);
@@ -123,20 +124,24 @@ function PositionItem ({ balance, claimPermissions, decimal, genesisHash, isSele
 
   return (
     <Motion variant='zoom'>
-      <Container disableGutters onClick={onClick} sx={{ alignItems: 'center', bgcolor: isSelected ? '#2D1E4A' : '#05091C', borderRadius: '14px', cursor: 'pointer', display: 'flex', flexDirection: 'row', justifyContent: 'space-between', p: '4px', pl: '18px' }}>
+      <Container disableGutters onClick={onClick} sx={{ alignItems: 'center', bgcolor: isSelected ? '#2D1E4A' : '#05091C', borderRadius: '14px', cursor: 'pointer', display: 'flex', flexDirection: 'row', justifyContent: 'space-between', p: totalPositions > 1 ? '4px' : '4px 30px 4px 4px', pl: '18px' }}>
         <TokenInfo genesisHash={genesisHash} />
         <Stack columnGap='5px' direction='row'>
           <StakingBadge hasPoolStaking={hasPoolStaking} isFullscreen />
-          {claimPermissionTooltip &&
-            <MyTooltip content={claimPermissionTooltip}>
-              <InfoCircle color='#674394' size={20} variant='Bulk' />
-            </MyTooltip>
-          }
+          <Box sx={{ width: '20px' }}>
+            {claimPermissionTooltip &&
+              <MyTooltip content={claimPermissionTooltip}>
+                <InfoCircle color='#674394' size={20} variant='Bulk' />
+              </MyTooltip>
+            }
+          </Box>
         </Stack>
         <TestnetBadge style={{ mt: 0, visibility: isTestNet ? 'visible' : 'hidden' }} />
         <ChainIdentifier genesisHash={genesisHash} />
         <Staked balance={balance} decimal={decimal} price={price} token={token} />
-        <ArrowButton onClick={noop} />
+        {totalPositions > 1 &&
+          <ArrowButton onClick={noop} />
+        }
       </Container>
     </Motion>
   );